requirements.txt
mysql-connector-python==8.1.0
aiohttp==3.8.5
aiogram==3.1.1
fpdf2==2.7.5
beautifulsoup4==4.12.2
lxml==4.9.3
aiofiles==23.2.1
